scrappy_web/requirements.txt
fastapi
uvicorn
python-multipart
python-jose[cryptography]
passlib[bcrypt]
slowapi
aiofiles
bcrypt
pytest
httpx
